backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
motor==3.7.1
pymongo==4.13.2
pydantic-settings==2.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
httpx==0.28.1
google-auth==2.38.0
email-validator==2.2.0
python-dotenv==1.0.1
